pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ivo"
version = "0.1.0"
description = "ivo (Intelligent Virtual Operator) — a Telegram chat front-end for LLMs. One bot, many models. Pluggable adapters for Copilot, Ollama, and beyond."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "ivo contributors" }]
keywords = ["telegram", "bot", "llm", "copilot", "ollama"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Communications :: Chat",
]
dependencies = [
  "python-telegram-bot>=21.0",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "kokoro-onnx>=0.4",
  "soundfile>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
]

[project.scripts]
ivo = "ivo.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ivo*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
